requirements.txt
amqp==5.0.6
aniso8601==8.1.1
asgiref==3.3.4
billiard==3.6.4.0
celery==5.1.1
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.3
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-graphql-jwt==0.3.2
django-prometheus==2.1.0
django-redis==4.12.1
graphene==3.0b7
graphene-crud-maker==0.0.6
graphene-django==3.0.0b7
graphql-core==3.1.4
graphql-relay==3.1.0
gunicorn==20.1.0
idna==2.10
kombu==5.1.0
Pillow==8.2.0
prometheus-client==0.10.1
promise==2.3
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
